backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Powered developer productivity platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic==0.7.7",
    "fastapi>=0.128.0",
    "google-genai>=1.4.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]
